skills/recruit-ops/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "recruit-ops"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary",
    "python-dateutil",
    "lark-oapi",
    "pdfminer-six==20200517",
]

[tool.setuptools]
packages = [
    "auto_reject",
    "common",
    "cron",
    "email_templates",
    "exam",
    "followup",
    "inbox",
    "intake",
    "interview",
    "lib",
    "lib.feishu",
    "ops",
    "outbound",
    "round1",
    "round2",
    "talent",
    "template",
    "tests",
]
py-modules = [
    "cron_health",
    "trigger_cron_now",
]

[tool.setuptools.package-dir]
"" = "scripts"
